스프링 부트 /springWebmvc/src/test/java/me/yjh/springWebmvc/user/UserControllerTest.java
package me.yjh.springWebmvc.user;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import static org.hamcrest.Matchers.equalTo;
import static org.hamcrest.Matchers.is;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@RunWith(SpringRunner.class)
@WebMvcTest(UserController.class)
public class UserControllerTest {

    @Autowired
    MockMvc mockMvc;

//    @Test
//    public void hello() throws Exception {
//
//        mockMvc.perform(get("/hello"))
//                .andExpect(status().isOk())
//                .andExpect(content().string("hello"));
//    }

    @Test
    public void createUser_JSON() throws Exception {

        String userJson = "{\"username\":\"jihoon\", \"password\":\"123\"}";

        mockMvc.perform(post("/users/create")
                .contentType(MediaType.APPLICATION_JSON_UTF8)
                .accept(MediaType.APPLICATION_JSON_UTF8)
                .content(userJson)) //여기까지는 요청을 만는 단계
                .andExpect(status().isOk())  //여기는 응답을 확인하는 단계
                .andExpect(jsonPath("$.username", is(equalTo("jihoon"))))
                .andExpect((jsonPath("$.password", is(equalTo("123")))));

    }


}
